refactored code into modules and moved demo code as well as fixed setup py file
