+++ labelsnow/constants.py
@@ -1,0 +1,17 @@
+LABELBOX_DEFAULT_TYPE_DICTIONARY = {
+    'ID': 'string',
+    'DataRow ID': 'string',
+    'Labeled Data': 'string',
+    'Created By': 'string',
+    'Project Name': 'string',
+    'Seconds to Label': 'float64',
+    'External ID': 'string',
+    'Agreement': 'int64',
+    'Benchmark Agreement': 'int64',
+    'Benchmark ID': 'string',
+    'Dataset Name': 'string',
+    'Reviews': 'object',
+    'View Label': 'string',
+    'Has Open Issues': 'int64',
+    'Skipped': 'bool'
+}
