+++ setup.py
@@ -1,15 +1,14 @@
-import setuptools
-
+import setuptools 
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
     name="labelsnow",
-    version=0.0.0,
+    version="0.0.1",
     author="Labelbox",
-    author_email="engineering@labelbox.com",
-    description="Labelbox Connector to Snowflake",
+    author_email="ecosystem+snowflake@labelbox.com",
+    description="Labelbox Connector for Snowflake",
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://labelbox.com",
